refactor(mapper-application): separar responsabilidades en creación de agregados
- Se refactorizaron los mappers de escritura (`AdministrativeReportWriteMapper`,`LedgerAccountWriteMapper`, `CompanyWriteMapper`, `ContractWriteMapper`, 
  `JournalEntryWriteMapper`, `OpeningBalanceWriteMapper`, `ThirdPartiesWriteMapper`, `AssignmentWriteMapper`).
- Se eliminaron invocaciones directas a métodos creacionales de agregados en los mappers.
- Cada mapper ahora se limita a transformar DTOs en objetos de dominio simples (VOs, IDs, enums, valores primitivos).
- Se actualizaron los Application Services correspondientes para invocar los métodos creacionales de los agregados 
  (`registerLedgerAccount`, `registerCompany`, `registerContract`, `registerJournalEntry`, `registerOpeningBalance`, 
  `registerThirdParties`, `assignPermanent`, `assignTemporary`) utilizando los mappers refactorizados.
- Resultado: mappers más simples y testables, Application Services responsables de la orquestación, y arquitectura alineada 
  con el ADR-(Aplicación)-01 Definición del contrato correcto de un Mapper en la capa de aplicación.

+++ src/main/java/com/example/ClinicaDefinitiva/application/dto/administration/contabilidad/contract/CreateContractDto.java
@@ -1,5 +1,6 @@
 package com.example.ClinicaDefinitiva.application.dto.administration.contabilidad.contract;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
@@ -13,5 +14,5 @@ public record CreateContractDto(
         String origin,
         LocalDate endDate,
         String coverageType,
-        Double coverageRate
+        BigDecimal coverageRate
 ) {}
